feat: replace condition to component mode
