+++ packages/xhear/src/main.mjs
@@ -3,7 +3,6 @@ import { meetsEle } from "./public.mjs";
 import { handler } from "./accessor.mjs";
 import renderFn from "./render/render.mjs";
 import fillFn from "./render/fill.mjs";
-import conditionFn from "./render/condition.mjs";
 import syncFn from "./render/sync.mjs";
 import eventFn from "./event.mjs";
 import LikeArray from "./array.mjs";
@@ -11,7 +10,7 @@ import formFn, { initFormEle } from "./form.mjs";
 import { getType, extend } from "../../stanz/src/public.mjs";
 import Stanz, { constructor } from "../../stanz/src/main.mjs";
 import watchFn from "../../stanz/src/watch.mjs";
-import "./render/condition-comp.mjs";
+import "./render/condition.mjs";
 const { defineProperties } = Object;
 
 const init = ({ _this, ele, proxySelf }) => {
@@ -249,8 +248,6 @@ export default class Xhear extends LikeArray {
 
 const sfn = Stanz.prototype;
 const fn = Xhear.prototype;
-
-fn.extend(conditionFn);
 
 fn.extend(
   {
