+++ packages/xhear/tests/statics/temp.html
@@ -17,16 +17,16 @@
     <div :else-if="num % 4 > 0">I am IF ELSE TWO div</div>
     <div :else>I am ELSE div</div> -->
     <x-if :value="num % 4 > 2">
-      <div>I am IF div {{num}}</div>
+      <div>I am IF div</div>
     </x-if>
     <x-else-if :value="num % 4 > 1">
-      <div>I am IF ELSE ONE div {{num}}</div>
+      <div>I am IF ELSE ONE div</div>
     </x-else-if>
     <x-else-if :value="num % 4 > 0">
-      <div>I am IF ELSE TWO div {{num}}</div>
+      <div>I am IF ELSE TWO div</div>
     </x-else-if>
     <x-else>
-      <div>I am ELSE div {{num}}</div>
+      <div>I am ELSE div</div>
     </x-else>
   </div>
   <br />
